fastapi_app/simple_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1 